+++ src/app/cli/create.go
@@ -32,16 +32,12 @@ func (opt *Create) Run(ctx app.Context) error {
 	if err != nil {
 		return err
 	}
-	return reconcile(
+	return applyReconciler(
 		opt.OutputFileArgs,
 		ctx,
-		func(pr *parser.ParseResult) (*parser.Reconciler, error) {
-			return parser.NewBlockReconciler(pr, func(r1 Record, r2 Record) bool {
-				return date.IsAfterOrEqual(r1.Date()) && r2.Date().IsAfterOrEqual(date)
-			})
-		},
-		func(r *parser.Reconciler) (Record, string, error) {
-			return r.AddBlock(lines)
+		func(pr *parser.ParseResult) (*parser.ReconcileResult, error) {
+			reconciler := parser.NewBlockReconciler(pr, date)
+			return reconciler.InsertBlock(lines)
 		},
 	)
 }