+++ src/parser/reconciler_test.go
@@ -1,7 +1,6 @@
 package parser
 
 import (
-	"errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	. "klog"
@@ -23,15 +22,14 @@ Hello World
     5h
 `
 	pr, _ := Parse(original)
-	reconciler, err := NewRecordReconciler(
-		pr,
-		nil,
-		func(r Record) bool { return r.Date().ToString() == "2018-01-02" },
-	)
+	reconciler := NewRecordReconciler(pr, func(r Record) bool {
+		return r.Date().ToString() == "2018-01-02"
+	})
+	require.NotNil(t, reconciler)
+	result, err := reconciler.AppendEntry(func(r Record) string { return "2h30m" })
 	require.Nil(t, err)
-	newRecord, reconciled, err := reconciler.AppendEntry(func(r Record) string { return "2h30m" })
-	require.Nil(t, err)
-	require.Equal(t, 150, newRecord.Entries()[2].Duration().InMinutes())
+	require.NotNil(t, result)
+	require.Equal(t, 150, result.NewRecord.Entries()[2].Duration().InMinutes())
 	assert.Equal(t, `
 2018-01-01
     1h
@@ -44,7 +42,7 @@ Hello World
 
 2018-01-03
     5h
-`, reconciled)
+`, result.NewText)
 }
 
 func TestReconcilerAddsNewlyCreatedEntryAtEndOfFile(t *testing.T) {
@@ -52,45 +50,33 @@ func TestReconcilerAddsNewlyCreatedEntryAtEndOfFile(t *testing.T) {
 2018-01-01
     1h`
 	pr, _ := Parse(original)
-	reconciler, err := NewRecordReconciler(
-		pr,
-		nil,
-		func(r Record) bool { return r.Date().ToString() == "2018-01-01" },
-	)
-	require.Nil(t, err)
-	_, reconciled, err := reconciler.AppendEntry(func(r Record) string { return "16:00-17:00" })
+	reconciler := NewRecordReconciler(pr, func(r Record) bool {
+		return r.Date().ToString() == "2018-01-01"
+	})
+	require.NotNil(t, reconciler)
+	result, err := reconciler.AppendEntry(func(r Record) string { return "16:00-17:00" })
 	require.Nil(t, err)
 	assert.Equal(t, `
 2018-01-01
     1h
     16:00-17:00
-`, reconciled)
+`, result.NewText)
 }
 
 func TestReconcilerSkipsIfNoRecordMatches(t *testing.T) {
 	original := "2018-01-01\n"
 	pr, _ := Parse(original)
-	reconciler, err := NewRecordReconciler(
-		pr,
-		errors.New("No such record"),
-		func(r Record) bool { return false },
-	)
+	reconciler := NewRecordReconciler(pr, func(r Record) bool { return false })
 	require.Nil(t, reconciler)
-	assert.EqualError(t, err, "No such record")
 }
 
 func TestReconcilerRejectsInvalidEntry(t *testing.T) {
 	original := "2018-01-01\n"
 	pr, _ := Parse(original)
-	reconciler, err := NewRecordReconciler(
-		pr,
-		errors.New("No such record"),
-		func(r Record) bool { return true },
-	)
-	require.Nil(t, err)
-	newRecord, reconciled, err := reconciler.AppendEntry(func(r Record) string { return "this is not valid entry text" })
-	assert.Equal(t, "", reconciled)
-	assert.Nil(t, newRecord)
+	reconciler := NewRecordReconciler(pr, func(r Record) bool { return true })
+	require.NotNil(t, reconciler)
+	result, err := reconciler.AppendEntry(func(r Record) string { return "this is not valid entry text" })
+	require.Nil(t, result)
 	assert.Error(t, err)
 }
 
@@ -102,20 +88,28 @@ func TestReconcilerClosesOpenRange(t *testing.T) {
 	2m
 `
 	pr, _ := Parse(original)
-	reconciler, err := NewRecordReconciler(
-		pr,
-		nil,
-		func(r Record) bool { return r.Date().ToString() == "2018-01-01" },
-	)
-	require.Nil(t, err)
-	_, reconciled, err := reconciler.CloseOpenRange(func(r Record) Time { return Ɀ_Time_(16, 42) })
+	reconciler := NewRecordReconciler(pr, func(r Record) bool {
+		return r.Date().ToString() == "2018-01-01"
+	})
+	require.NotNil(t, reconciler)
+	result, err := reconciler.CloseOpenRange(func(r Record) Time { return Ɀ_Time_(16, 42) })
 	require.Nil(t, err)
 	assert.Equal(t, `
 2018-01-01
     1h
     15:00-16:42 Will this close? I hope so!?!?
 	2m
-`, reconciled)
+`, result.NewText)
+}
+
+func TestReconcileAddBlockIfOriginalIsEmpty(t *testing.T) {
+	pr, _ := Parse("")
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(3333, 1, 1))
+	result, err := reconciler.InsertBlock([]parsing.Text{
+		{"2000-05-05", 0},
+	})
+	require.Nil(t, err)
+	assert.Equal(t, "2000-05-05\n", result.NewText)
 }
 
 func TestReconcileAddBlockToEnd(t *testing.T) {
@@ -123,11 +117,8 @@ func TestReconcileAddBlockToEnd(t *testing.T) {
 2018-01-01
     1h`
 	pr, _ := Parse(original)
-	reconciler, _ := NewBlockReconciler(
-		pr,
-		func(Record, Record) bool { return false },
-	)
-	_, reconciled, err := reconciler.AddBlock([]parsing.Text{
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(2018, 1, 2))
+	result, err := reconciler.InsertBlock([]parsing.Text{
 		{"2018-01-02", 0},
 	})
 	require.Nil(t, err)
@@ -136,7 +127,56 @@ func TestReconcileAddBlockToEnd(t *testing.T) {
     1h
 
 2018-01-02
-`, reconciled)
+`, result.NewText)
+}
+
+func TestReconcileAddBlockToEndWithTrailingNewlines(t *testing.T) {
+	original := `
+2018-01-01
+    1h
+
+`
+	pr, _ := Parse(original)
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(2018, 1, 2))
+	result, err := reconciler.InsertBlock([]parsing.Text{
+		{"2018-01-02", 0},
+	})
+	require.Nil(t, err)
+	assert.Equal(t, `
+2018-01-01
+    1h
+
+2018-01-02
+
+`, result.NewText)
+}
+
+func TestReconcileAddBlockToBeginning(t *testing.T) {
+	original := "2018-01-02"
+	pr, _ := Parse(original)
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(2018, 1, 1))
+	result, err := reconciler.InsertBlock([]parsing.Text{
+		{"2018-01-01", 0},
+	})
+	require.Nil(t, err)
+	assert.Equal(t, `2018-01-01
+
+2018-01-02`, result.NewText)
+}
+
+func TestReconcileAddBlockToBeginningWithLeadingNewlines(t *testing.T) {
+	original := "\n\n2018-01-02"
+	pr, _ := Parse(original)
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(2018, 1, 1))
+	result, err := reconciler.InsertBlock([]parsing.Text{
+		{"2018-01-01", 0},
+	})
+	require.Nil(t, err)
+	assert.Equal(t, `2018-01-01
+
+
+
+2018-01-02`, result.NewText)
 }
 
 func TestReconcileAddBlockInBetween(t *testing.T) {
@@ -147,11 +187,8 @@ func TestReconcileAddBlockInBetween(t *testing.T) {
 2018-01-03
     3h`
 	pr, _ := Parse(original)
-	date := Ɀ_Date_(2018, 1, 2)
-	reconciler, _ := NewBlockReconciler(pr, func(r1 Record, r2 Record) bool {
-		return date.IsAfterOrEqual(r1.Date()) && r2.Date().IsAfterOrEqual(date)
-	})
-	_, reconciled, err := reconciler.AddBlock([]parsing.Text{
+	reconciler := NewBlockReconciler(pr, Ɀ_Date_(2018, 1, 2))
+	result, err := reconciler.InsertBlock([]parsing.Text{
 		{"2018-01-02", 0},
 		{"This and that", 0},
 		{"30m worked", 1},
@@ -166,5 +203,5 @@ This and that
     30m worked
 
 2018-01-03
-    3h`, reconciled)
+    3h`, result.NewText)
 }